+++ commonjs/jwt-http.service.d.ts
@@ -1,6 +1,7 @@
 import { Http, Response, RequestOptionsArgs, Request } from '@angular/http';
 import { Observable } from 'rxjs/Observable';
 import 'rxjs/add/operator/switchMap';
+import 'rxjs/add/operator/catch';
 import { ConfigService } from './config.service';
 import { SharedService } from './shared.service';
 export interface JwtRequestOptionsArgs extends RequestOptionsArgs {